pyproject.toml
[project]
name = "the-house-reloaded"
version = "0.1.0"
description = "A python rewrite of the original The House board in PHP."
authors = [{name = "Husayn Haras", email = "haras@disroot.org"}]
maintainers = [{name = "Husayn Haras", email = "haras@disroot.org"}]
readme = "README.md"
license = "GPL-3.0-only"
license-files = ["LICENSE"]
requires-python = ">=3.12"
dependencies = [
    "bleach>=6.2.0",
    "flask>=3.1.0",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-marshmallow>=1.3.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "marshmallow-sqlalchemy>=1.4.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9.10",
]
